requirements.txt
pytorch
pytorch_pretrained_bert